# docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Work Flow API
  version: 0.0.1
paths:
...
  /app/create:
    post:
      parameters:
      - description: 创建人
        in: body
        name: data
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建表单成功"}'
          schema:
            type: string
      summary: 创建表单
      tags:
      - App
  /app/empty:
    get:
      parameters:
      - description: 应用id
        in: query
        name: appId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"null"}'
          schema:
            type: string
      summary: 返回空表单
      tags:
      - App
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"